requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cryptography==40.0.2
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
distro==1.8.0
Django==4.2
django-allauth==0.54.0
django-appconf==1.0.5
django-environ==0.10.0
django-responsive2==0.1.3
django-social-share==2.3.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
fonttools==4.45.1
fqdn==1.5.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
ipykernel==6.27.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
oauthlib==3.2.2
openai==0.28.0
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandasql==0.7.3
pandocfilters==1.5.0
parso==0.8.3
Pillow==9.5.0
platformdirs==4.0.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psutil==5.9.6
psycopg2==2.9.6
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
PyJWT==2.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
razorpay==1.3.0
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.1
seaborn==0.13.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.3
stack-data==0.6.3
terminado==0.18.0
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.15
virtualenv==20.25.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
xlrd==2.0.1
yarl==1.9.4
